app/LooseTheLeftovers_Backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-admin-cli==0.1.1
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
black==23.9.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1
Django-crontab==0.7.1
polars==0.20.16
django-rest-passwordreset==1.4.0
